+++ python/actorcore/TestHelper.py
@@ -137,12 +137,14 @@ noneLoaded = {'cartridgeLoaded':[-1,-1,'?',-1,-1]}
 bossLoaded = {'cartridgeLoaded':[11,1000,'A',54321,1], 'survey':['BOSS','None'], 'loadedNewCartridge':[]}
 apogeeLoaded = {'cartridgeLoaded':[1,2000,'A',54321,2], 'survey':['APOGEE','None'], 'loadedNewCartridge':[]}
 mangaDitherLoaded = {'cartridgeLoaded':[2,3000,'A',54321,3], 'survey':['MaNGA','MaNGA Dither'], 'loadedNewCartridge':[]}
+mangaGlobularLoaded = {'cartridgeLoaded':[2,3000,'A',54321,3], 'survey':['MaNGA','MaNGA Globular'], 'loadedNewCartridge':[]}
 manga10Loaded = {'cartridgeLoaded': [2, 3000, 'A', 54321, 3], 'survey': ['MaNGA', 'MaNGA 10min'],
                  'loadedNewCartridge': []}
 mangaStareLoaded = {'cartridgeLoaded':[2,3001,'A',54321,3], 'survey':['MaNGA','MaNGA Stare'], 'loadedNewCartridge':[]}
 MaStarLoaded = {'cartridgeLoaded': [2, 3001, 'A', 54321, 3],
                 'survey': ['MaNGA', 'MaStar'], 'loadedNewCartridge': []}
 apogeemangaDitherLoaded = {'cartridgeLoaded':[3,4000,'A',54321,4], 'survey':['APOGEE-2&MaNGA','MaNGA Dither'], 'loadedNewCartridge':[]}
+apogeemangaGlobularLoaded = {'cartridgeLoaded':[3,4000,'A',54321,4], 'survey':['APOGEE-2&MaNGA','MaNGA Globular'], 'loadedNewCartridge':[]}
 apogeemangaStareLoaded = {'cartridgeLoaded':[3,4000,'A',54321,4], 'survey':['APOGEE-2&MaNGA','MaNGA Stare'], 'loadedNewCartridge':[]}
 apogeemangaMaStarLoaded = {'cartridgeLoaded': [3, 4000, 'A', 54321, 4],
                            'survey': ['APOGEE-2&MaNGA', 'MaStar'], 'loadedNewCartridge': []}
@@ -173,10 +175,11 @@ guiderState['noneLoaded'] = merge_dicts(guiderOff,noneLoaded,mangaC)
 guiderState['cartLoaded'] = merge_dicts(guiderOff,bossLoaded,mangaC)
 guiderState['bossLoaded'] = merge_dicts(guiderOff,bossLoaded,mangaC)
 guiderState['apogeeLoaded'] = merge_dicts(guiderOff,apogeeLoaded,mangaC)
-guiderState['mangaStareLoaded'] = merge_dicts(guiderOff,mangaDitherLoaded,mangaC)
+guiderState['mangaDitherLoaded'] = merge_dicts(guiderOff,mangaDitherLoaded,mangaC)
 guiderState['manga10Loaded'] = merge_dicts(guiderOff, manga10Loaded, mangaC)
 guiderState['MaStarLoaded'] = merge_dicts(guiderOff, MaStarLoaded, mangaC)
-guiderState['mangaDitherLoaded'] = merge_dicts(guiderOff,mangaStareLoaded,mangaC)
+guiderState['mangaStareLoaded'] = merge_dicts(guiderOff,mangaStareLoaded,mangaC)
+guiderState['mangaGlobularLoaded'] = merge_dicts(guiderOff,mangaGlobularLoaded,mangaC)
 guiderState['apogeemangaDitherLoaded'] = merge_dicts(guiderOff,apogeemangaDitherLoaded,mangaC)
 guiderState['apogeemanga10Loaded'] = merge_dicts(guiderOff, apogeemanga10Loaded, mangaC)
 guiderState['apogeemangaNoneLoaded'] = merge_dicts(guiderOff,
@@ -184,6 +187,7 @@ guiderState['apogeemangaNoneLoaded'] = merge_dicts(guiderOff,
                                                    mangaC)
 guiderState['apogeemangaStareLoaded'] = merge_dicts(guiderOff,apogeemangaStareLoaded,mangaC)
 guiderState['apogeemangaMaStarLoaded'] = merge_dicts(guiderOff, apogeemangaMaStarLoaded, mangaC)
+guiderState['apogeemangaGlobularLoaded'] = merge_dicts(guiderOff,apogeemangaGlobularLoaded,mangaC)
 guiderState['apogeeLeadLoaded'] = merge_dicts(guiderOff,apogeeLeadLoaded,mangaC)
 guiderState['guiderOn'] = merge_dicts(guiderOn,bossLoaded,mangaC)
 guiderState['guiderOnDecenter'] = merge_dicts(guiderOn,mangaDitherLoaded,mangaN)
@@ -193,12 +197,14 @@ guiderState['guiderOnDecenter'] = merge_dicts(guiderOn,mangaDitherLoaded,mangaN)
 bossPointing = {'pointingInfo':[1000,11,'A',10.,20.,1.,0.,5500,'BOSS','BOSS']}
 apogeePointing = {'pointingInfo':[2000,1,'A',20.,30.,2.,1.,10500,'APOGEE-2','APOGEE']}
 mangaDitherPointing = {'pointingInfo':[3000,2,'A',30.,40.,3.,2.,5500,'MaNGA','MaNGA dither']}
+mangaGlobularPointing = {'pointingInfo': [3000, 2, 'A', 30., 40., 3., 2., 5500, 'MaNGA', 'MaNGA Globular']}
 manga10Pointing = {'pointingInfo': [3001, 2, 'A', 30., 40., 3., 2., 5500, 'MaNGA', 'MaNGA 10min']}
 mangaStarePointing = {'pointingInfo':[3001,2,'A',30.,40.,3.,2.,5500,'MaNGA','MaNGA Stare']}
 MaStarPointing = {'pointingInfo': [3001, 2, 'A', 30., 40., 3., 2., 5500, 'MaNGA', 'MaStar']}
 apogeeLeadPointing = {'pointingInfo':[4000,3,'A',40.,50.,4.,3.,10500,'APOGEE2&MANGA','APOGEE lead']}
 apogeeLeadPointingCart7 = {'pointingInfo':[4000,7,'A',40.,50.,4.,3.,10500,'APOGEE2&MANGA','APOGEE lead']}
 apgoeemangaDitherPointing = {'pointingInfo':[5000,4,'A',50.,60.,5.,4.,5500,'APOGEE-2&MaNGA','MaNGA dither']}
+apgoeemangaGlobularPointing = {'pointingInfo': [5000, 4, 'A', 50., 60., 5., 4., 5500, 'APOGEE-2&MaNGA', 'MaNGA Globular']}
 apgoeemanga10Pointing = {'pointingInfo':[5000,4,'A',50.,60.,5.,4.,5500,'APOGEE-2&MaNGA','MaNGA 10min']}
 apogeemangaStarePointing = {'pointingInfo':[5001,4,'A',50.,60.,5.,4.,5500,'APOGEE-2&MaNGA','MaNGA Stare']}
 apogeemangaMaStarPointing = {'pointingInfo': [5001, 4, 'A', 50., 60., 5., 4., 5500,
@@ -217,10 +223,12 @@ platedbState['apogee'] = merge_dicts(apogeePointing,apogeeDesignNone, APOGEEInst
 platedbState['mangaDither'] = merge_dicts(mangaDitherPointing,apogeeDesignNone, BOSSInstrumentPlugged)
 platedbState['manga10'] = merge_dicts(manga10Pointing,apogeeDesignNone, BOSSInstrumentPlugged)
 platedbState['mangaStare'] = merge_dicts(mangaStarePointing,apogeeDesignNone, BOSSInstrumentPlugged)
+platedbState['mangaGlobular'] = merge_dicts(mangaGlobularPointing, apogeeDesignNone, BOSSInstrumentPlugged)
 platedbState['MaStar'] = merge_dicts(MaStarPointing, apogeeDesignNone, BOSSInstrumentPlugged)
 platedbState['MaStar_short'] = merge_dicts(MaStarPointing, apogeeDesignNone_manga_short, BOSSInstrumentPlugged)
 platedbState['MaStar_coobs_short'] = merge_dicts(MaStarPointing, apogeeDesignNone_manga_short, BothInstrumentsPlugged)
 platedbState['apgoeemangaDither'] = merge_dicts(apgoeemangaDitherPointing,apogeeDesignNone, BothInstrumentsPlugged)
+platedbState['apogeemangaGlobular'] = merge_dicts(apgoeemangaGlobularPointing, apogeeDesignNone, BothInstrumentsPlugged)
 platedbState['apgoeemanga10'] = merge_dicts(apgoeemanga10Pointing,apogeeDesignNone, BothInstrumentsPlugged)
 platedbState['apgoeemangaStare'] = merge_dicts(apogeemangaStarePointing,apogeeDesignNone, BothInstrumentsPlugged)
 platedbState['apogeemangaMaStar'] = merge_dicts(apogeemangaMaStarPointing, apogeeDesignNone,
@@ -255,8 +263,8 @@ ecameraState['ok'] = merge_dicts(gcameraTempOk,gcameraSeqNo,gcameraSimulatingOff
 # sop state setup
 bypasses = ["ffs", "lamp_ff", "lamp_hgcd", "lamp_ne", "axes",
             "isBoss", "isApogee",
-            "isMangaDither", "isMangaStare", 'isManga10', 'isApogeeManga10',
-            "isApogeeLead", "isApogeeMangaDither", "isApogeeMangaStare",
+            "isMangaDither", "isMangaStare", 'isManga10', 'isMangaGlobular' 'isApogeeManga10',
+            "isApogeeLead", "isApogeeMangaDither", "isApogeeMangaStare", 'isApogeeMangaGlobular',
             "gangToCart", "gangToPodium", "slewToField", "guiderDark", "guider_decenter",
             'isMaStar', 'isApogeeMangaMaStar']
 sopNoBypass = {'bypassNames':bypasses, 'bypassedNames':[]}